treesitter_code_analyzer/pyproject.toml
[project]
name = "treesitter-code-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
    "tree-sitter>=0.25.0",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-javascript>=0.20.4",
    "uvicorn>=0.35.0",
    "graphviz>=0.21",
]


[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
